% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famdviz.R
\name{get_variables}
\alias{get_variables}
\title{Get Variables}
\usage{
get_variables(famd, dim_x, dim_y, color_by)
}
\arguments{
\item{famd}{list produced by factoMineR::FAMD function}

\item{dim_x}{first dimension integer}

\item{dim_y}{second dimension integer}

\item{color_by}{"contrib" or "cos2" (see factoMineR::FAMD documentation)}
}
\value{
a data frame of x, y coordinates and color intensities
}
\description{
Get variables of factoMineR::FAMD model as an input for plot_variables function
}
\examples{
famd <- factoMineR::FAMD(x, graph = F)
get_variables(famd, 1, 2, "cos2")
}
