% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famdviz.R
\name{plot_variables}
\alias{plot_variables}
\title{Plot variables}
\usage{
plot_variables(famd, dim_x, dim_y, color_by)
}
\arguments{
\item{x}{the numeric value}
}
\value{
the square root of x
}
\description{
Calculate the square root of a numeric value
}
\examples{
sqroot(22);
}
