% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famdviz.R
\name{set_custom_theme}
\alias{set_custom_theme}
\title{Set Custom Theme}
\usage{
set_custom_theme(
  legend_position_x,
  legend_position_y,
  axis_text_x = T,
  axis_text_y = T
)
}
\arguments{
\item{legend_position_x}{numeric value between 0 and 1 indicating left-right position of legend}

\item{legend_position_y}{numeric value between 0 and 1 indicating top-down position of legend}

\item{axis_text_x}{boolean value indicating labels on horizontal axis tics}

\item{axis_text_y}{boolean value indicating labels on vertical axis tics}
}
\value{
ggplot2::theme()
}
\description{
Supporting function for ggplot generating customized theme
}
\examples{
data_frame <- data.frame(x = c(1, 2, 3),
                         y = c(1, 2, 3),
                         z = c(12.8, 6.5, 15.4))
mean_color <- mean(data_frame[, "z"])
ggplot(data_frame, aes(x = x, y = y, color = z) +
  geom_point(shape = 17) +
  set_gradient(mean_color)
}
