% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famdviz.R
\name{set_gradient}
\alias{set_gradient}
\title{Set Gradient}
\usage{
set_gradient(midpoint)
}
\arguments{
\item{midpoint}{the numeric value indicating mid color}
}
\value{
scale_colour_gradient2()
}
\description{
Supporting function for ggplot generates 3-point scale_colour_gradient2
}
\examples{
data_frame <- data.frame(x = c(1, 2, 3),
                         y = c(1, 2, 3),
                         z = c(12.8, 6.5, 15.4))
mean_color <- mean(data_frame[, "z"])
ggplot(data_frame, aes(x = x, y = y, color = z) +
  geom_point(shape = 17) +
  set_gradient(mean_color)
}
